+++ changelog.md
@@ -6,6 +6,7 @@
 - allow lock the dock/toolbar items.
 - show instances counter of the same app on dock.
 - allow set the toolbar on different positions.
+- allow set custom start menu icon.
 
 ### enhancements
 - settings shown each time on startup.